Switch E3SM-Unified deployment to use mache
Information about E3SM supported machines now comes from mache,
rather than from config files within the E3SM-Unified repo.  This
should help to maintain compatibility between E3SM-Unified and
other analysis tools.

Since bootstrapping relies on having mache installed in a temporary
conda environment, the deployment script has been broken into 3
parts.  The `deploy_e3sm_unified.py` script now first creates the
installation conda environment, then calls the `bootstrap.py`
script using that environment.  Any functions shared between these
two scripts are in `shared.py`.

A new `$E3SMU_MACHINE` environment variable has been added that
stores the machine name for later identification (e.g. on a
compute node).
