+++ e3sm_supported_machines/default.cfg
@@ -1,6 +1,6 @@
 # Options related to deploying a e3sm-unified conda environment on supported
 # machines
-[deploy]
+[e3sm_unified]
 
 # is this a shared production environment?
 release = False
@@ -28,3 +28,7 @@ tempest_extremes = 2.2
 
 # the version of ILAMB to build if using system compilers and MPI
 ilamb = 2.6
+
+# the version of mache to use during deployment (should match the version used
+# in the package itself)
+mache = 1.0.0
