--- e3sm_supported_machines/machines/acme1.cfg
@@ -1,11 +1,0 @@
-# Options related to deploying an e3sm-unified conda environment on supported
-# machines
-[deploy]
-
-# the unix group for permissions for the e3sm-unified conda environment
-group = climate
-
-# the path to the directory where activation scripts, the base environment, and
-# system libraries will be deployed
-base_path = /usr/local/e3sm_unified/envs
-
