--- e3sm_supported_machines/machines/anvil.cfg
@@ -1,16 +1,0 @@
-# Options related to deploying an e3sm-unified conda environment on supported
-# machines
-[deploy]
-
-# the unix group for permissions for the e3sm-unified conda environment
-group = cels
-
-# the compiler set to use for system libraries and MPAS builds
-compiler = intel
-
-# the system MPI library to use for intel18 compiler
-mpi = impi
-
-# the path to the directory where activation scripts, the base environment, and
-# system libraries will be deployed
-base_path = /lcrc/soft/climate/e3sm-unified
