+++ e3sm_supported_machines/templates/load_e3sm_unified.csh.template
@@ -1,6 +1,7 @@
 source {{ conda_base }}/etc/profile.d/conda.csh
 
 setenv E3SMU_SCRIPT "{{ script_filename }}"
+setenv E3SMU_MACHINE "{{ machine }}"
 
 if ( ! $?SLURM_JOB_ID && ! $?COBALT_JOBID ) then
   # we seem to be on a login node, so load the no-MPI environment
