+++ recipes/e3sm-unified/meta.yaml
@@ -1,5 +1,5 @@
 {% set name = "E3SM-Unified" %}
-{% set version = "1.5.0" %}
+{% set version = "1.6.0" %}
 {% set build = 0 %}
 
 package:
@@ -36,6 +36,7 @@ requirements:
     - ipython
     - jupyter
     - livvkit 3.0.0
+    - mache 1.0.0
     - moab 5.3.0
     - mpas-analysis 1.4.0
     - mpas_tools 0.7.0
